move indexed aead keys to types crate
