+++ coreclient/src/clients/add_contact.rs
@@ -7,7 +7,8 @@ use openmls::group::GroupId;
 use phnxtypes::{
     codec::PhnxCodec,
     crypto::{
-        ear::keys::FriendshipPackageEarKey, hpke::HpkeEncryptable, signatures::signable::Signable,
+        ear::keys::FriendshipPackageEarKey, hpke::HpkeEncryptable,
+        indexed_aead::keys::UserProfileKey, signatures::signable::Signable,
     },
     identifiers::{Fqdn, QsReference, QualifiedUserName},
     messages::{
@@ -23,7 +24,9 @@ use crate::{
     Conversation, ConversationAttributes, ConversationId, PartialContact, UserProfile,
     clients::connection_establishment::{ConnectionEstablishmentPackageTbs, FriendshipPackage},
     groups::{Group, PartialCreateGroupParams, openmls_provider::PhnxOpenMlsProvider},
-    key_stores::{MemoryUserKeyStore, as_credentials::AsCredentials, indexed_keys::UserProfileKey},
+    key_stores::{
+        MemoryUserKeyStore, as_credentials::AsCredentials, indexed_keys::StorableIndexedKey,
+    },
     store::StoreNotifier,
 };
