+++ types/src/crypto/indexed_aead/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2025 Phoenix R&D GmbH <hello@phnx.im>
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+pub mod keys;
